erasure-code: isa/lrc plugin feature
There are two new plugins (isa and lrc). When upgrading a cluster, there
must be a protection against the following scenario:

    * the mon are upgraded but not the osd
    * a new pool is created using plugin isa
    * the osd fail to load the isa plugin because they have not been
      upgraded

A feature bit is added : PLUGINS_V2. The monitor will only agree to
create an erasure code profile for the isa or lrc plugin if all OSDs
supports PLUGINS_V2. Once such an erasure code profile is stored in the
OSDMap, an OSD can only boot if it supports the PLUGINS_V2 feature,
which means it is able to load the isa and lrc plugins.

This is going to be needed each time new plugins are added, which is
impractical. A more generic plugin upgrade support should be added
instead, as described in http://tracker.ceph.com/issues/7291.

http://tracker.ceph.com/issues/9343 Refs: #9343

Signed-off-by: Loic Dachary <loic-201408@dachary.org>

s