mon: --mon_debug_idempotency repeats OSD commands
The mon command is run twice to check for idempotency and cover the case
where the paxos proposal with the pending incremental changes have not
yet been submitted. For instance:

   ceph tell 'mon.*' injectargs -- --mon-debug-idempotency
   ceph osd pool create mypool 12
   ceph tell 'mon.*' injectargs -- --no-mon-debug-idempotency

will create mypool twice.

This is equivalent to CEPH_CLI_TEST_DUP_COMMAND in ceph.in except it is
done within the monitor and is not subject to race conditions where the
second command arrives after the first command changes have been
submitted.

Signed-off-by: Loic Dachary <loic-201408@dachary.org>