doc: Changed Crush mapper temporary buffer length
Crush temporary buffers are allocated as per replica size configured
by the user.When there are more final osds (to be selected as per
rule) than the replicas, buffer overlaps and it causes crash. Hence
temporary buffer is allocated to max_size parameter that specifies
the maximum osds that can be selected using the rule.

Fixes: #9492

Signed-off-by: Johnu George <johnugeo@cisco.com>